+++ test/transforms/test_mutual_information_sort.py
@@ -1,3 +1,4 @@
+import pytest
 import torch
 
 from torch_frame import TaskType, TensorFrame, stype
@@ -6,10 +7,13 @@ from torch_frame.datasets.fake import FakeDataset
 from torch_frame.transforms import MutualInformationSort
 
 
-def test_mutual_information_sort_classification():
-    dataset: Dataset = FakeDataset(
-        num_rows=10, with_nan=False, stypes=[stype.numerical],
-        create_split=True, task_type=TaskType.MULTICLASS_CLASSIFICATION)
+@pytest.mark.parametrize('with_nan', [True, False])
+@pytest.mark.parametrize(
+    'task_type', [TaskType.MULTICLASS_CLASSIFICATION, TaskType.REGRESSION])
+def test_mutual_information_sort(with_nan, task_type):
+    dataset: Dataset = FakeDataset(num_rows=10, with_nan=with_nan,
+                                   stypes=[stype.numerical], create_split=True,
+                                   task_type=task_type)
     # modify the FakeDataset so column c would have highest mutual information
     # score
     dataset.df['c'] = dataset.df['target'].astype(float)
@@ -18,47 +22,28 @@ def test_mutual_information_sort_classification():
 
     tensor_frame: TensorFrame = dataset.tensor_frame
     train_dataset = dataset.get_split_dataset('train')
-    transform = MutualInformationSort(TaskType.MULTICLASS_CLASSIFICATION)
+    transform = MutualInformationSort(task_type)
     transform.fit(train_dataset.tensor_frame)
     out = transform(tensor_frame)
 
     # column c ranks the first
     assert (out.col_names_dict[stype.numerical][0] == 'c')
-    assert (torch.eq(out.x_dict[stype.numerical][:, 0],
-                     torch.tensor(dataset.df['c'].values,
-                                  dtype=torch.float32)).all())
+    actual_first_col = out.x_dict[stype.numerical][:, 0]
+    actual_first_col_nan_mask = torch.isnan(actual_first_col)
+    expected_first_col = torch.tensor(dataset.df['c'].values,
+                                      dtype=torch.float32)
+    expected_first_col_nan_mask = torch.isnan(expected_first_col)
+    # if the tensor on first column contains NaNs, make sure the NaNs
+    # are unchanged
+    assert (torch.allclose(actual_first_col_nan_mask,
+                           expected_first_col_nan_mask))
+    actual = actual_first_col[~actual_first_col_nan_mask]
+    expected = expected_first_col[~expected_first_col_nan_mask]
+    # make sure that the non NaN values are the same on first column
+    assert (torch.allclose(actual, expected))
 
     # make sure the shapes are unchanged
     assert (set(out.col_names_dict[stype.numerical]) == set(
         tensor_frame.col_names_dict[stype.numerical]))
-    assert (out.x_dict[stype.numerical].size() == tensor_frame.x_dict[
-        stype.numerical].size())
-
-
-def test_mutual_information_sort_regression():
-    dataset: Dataset = FakeDataset(num_rows=10, with_nan=False,
-                                   stypes=[stype.numerical], create_split=True,
-                                   task_type=TaskType.REGRESSION)
-    # modify the FakeDataset so column c would have highest mutual information
-    # score
-    dataset.df['c'] = dataset.df['target'].astype(float)
-    dataset.materialize()
-
-    tensor_frame: TensorFrame = dataset.tensor_frame
-    train_dataset = dataset.get_split_dataset('train')
-    transform = MutualInformationSort(TaskType.REGRESSION)
-    transform.fit(train_dataset.tensor_frame)
-    out = transform(tensor_frame)
-
-    # column c ranks the first
-    assert (out.col_names_dict[stype.numerical][0] == 'c')
-    assert (torch.allclose(
-        out.x_dict[stype.numerical][:, 0],
-        torch.tensor(dataset.df['c'].values, dtype=torch.float32)))
-
-    # make sure the column names are unchanged
-    assert (set(out.col_names_dict[stype.numerical]) == set(
-        tensor_frame.col_names_dict[stype.numerical]))
-    # make sure the tensor shapes are unchanged
     assert (out.x_dict[stype.numerical].size() == tensor_frame.x_dict[
         stype.numerical].size())